fix: ajusta migrations
